# Fix crash during stream defrag