This crash was introduced by https://github.com/redis/redis/issues/14130, which omitted the update of consume group in `s->cgroups_ref`.

In this PR, when performing defragmentation on `s->cgroups` using `defragRadixTree()`, we no longer rely on the automatic data defragmentation of `defragRadixTree()`. Instead, we manually defragment the consumer group and then update its reference in `s->cgroups`.

crash report:
```
Logged crash report (pid 117196):
=== REDIS BUG REPORT START: Cut & paste starting from here ===
117196:M 01 Sep 2025 18:48:19.908 # === ASSERTION FAILED ===
117196:M 01 Sep 2025 18:48:19.908 # ==> t_stream.c:2662 'raxFind(group->pel, buf, sizeof(buf), (void **)&nack)' is not true

------ STACK TRACE ------

117199 bio_aof
/lib/x86_64-linux-gnu/libc.so.6(+0x98d71)[0x7f05a6e98d71]
/lib/x86_64-linux-gnu/libc.so.6(pthread_cond_wait+0x20d)[0x7f05a6e9b7ed]
src/redis-server 127.0.0.1:21123(bioProcessBackgroundJobs+0x2b7)[0x558e45c81ff7]
/lib/x86_64-linux-gnu/libc.so.6(+0x9caa4)[0x7f05a6e9caa4]
/lib/x86_64-linux-gnu/libc.so.6(+0x129c3c)[0x7f05a6f29c3c]

117198 bio_close_file
/lib/x86_64-linux-gnu/libc.so.6(+0x98d71)[0x7f05a6e98d71]
/lib/x86_64-linux-gnu/libc.so.6(pthread_cond_wait+0x20d)[0x7f05a6e9b7ed]
src/redis-server 127.0.0.1:21123(bioProcessBackgroundJobs+0x2b7)[0x558e45c81ff7]
/lib/x86_64-linux-gnu/libc.so.6(+0x9caa4)[0x7f05a6e9caa4]
/lib/x86_64-linux-gnu/libc.so.6(+0x129c3c)[0x7f05a6f29c3c]

117196 redis-server *
src/redis-server 127.0.0.1:21123(streamCleanupEntryCGroupRefs+0x1b8)[0x558e45cf26d8]
src/redis-server 127.0.0.1:21123(xdelexCommand+0x3fb)[0x558e45cfb56b]
src/redis-server 127.0.0.1:21123(call+0x171)[0x558e45b98aa1]
src/redis-server 127.0.0.1:21123(processCommand+0xae8)[0x558e45b9a198]
src/redis-server 127.0.0.1:21123(processInputBuffer+0xd9)[0x558e45bc0a89]
src/redis-server 127.0.0.1:21123(readQueryFromClient+0x358)[0x558e45bc6438]
src/redis-server 127.0.0.1:21123(+0x226754)[0x558e45d0e754]
src/redis-server 127.0.0.1:21123(aeMain+0xf9)[0x558e45b7b9e9]
src/redis-server 127.0.0.1:21123(main+0x4a7)[0x558e45b758d7]
/lib/x86_64-linux-gnu/libc.so.6(+0x2a1ca)[0x7f05a6e2a1ca]
/lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0x8b)[0x7f05a6e2a28b]
src/redis-server 127.0.0.1:21123(_start+0x25)[0x558e45b77155]

117200 bio_lazy_free
/lib/x86_64-linux-gnu/libc.so.6(+0x98d71)[0x7f05a6e98d71]
/lib/x86_64-linux-gnu/libc.so.6(pthread_cond_wait+0x20d)[0x7f05a6e9b7ed]
src/redis-server 127.0.0.1:21123(bioProcessBackgroundJobs+0x2b7)[0x558e45c81ff7]
/lib/x86_64-linux-gnu/libc.so.6(+0x9caa4)[0x7f05a6e9caa4]
/lib/x86_64-linux-gnu/libc.so.6(+0x129c3c)[0x7f05a6f29c3c]

4/4 expected stacktraces.

------ STACK TRACE DONE ------

------ INFO OUTPUT ------
# Server
redis_version:255.255.255
redis_git_sha1:5f8e7852
redis_git_dirty:1
redis_build_id:d92741c8ea377c9d
redis_mode:standalone
os:Linux 6.8.0-79-generic x86_64
arch_bits:64
monotonic_clock:POSIX clock_gettime
multiplexing_api:epoll
atomicvar_api:c11-builtin
gcc_version:13.3.0
process_id:117196
process_supervised:no
run_id:da0b7541a2b521bb624ec0204efa5e26a80c6cdf
tcp_port:21123
server_time_usec:1756723699908962
uptime_in_seconds:0
uptime_in_days:0
hz:10
configured_hz:10
lru_clock:11893235
executable:/home/sundb/data/rf_1/src/redis-server
config_file:/home/sundb/data/rf_1/./tests/tmp/redis.conf.116679.26
io_threads_active:0
listener0:name=tcp,bind=127.0.0.1,port=21123
listener1:name=unix,bind=/home/sundb/data/rf_1/tests/tmp/server.116679.25/socket

# Clients
connected_clients:1
cluster_connections:0
maxclients:10000
client_recent_max_input_buffer:0
client_recent_max_output_buffer:0
blocked_clients:0
tracking_clients:0
pubsub_clients:0
watching_clients:0
clients_in_timeout_table:0
total_watched_keys:0
total_blocking_keys:0
total_blocking_keys_on_nokey:0

# Memory
used_memory:884344
used_memory_human:863.62K
used_memory_rss:9961472
used_memory_rss_human:9.50M
used_memory_peak:903272
used_memory_peak_human:882.10K
used_memory_peak_time:1756723699
used_memory_peak_perc:97.90%
used_memory_overhead:720816
used_memory_startup:652096
used_memory_dataset:163528
used_memory_dataset_perc:70.41%
allocator_allocated:1160848
allocator_active:1503232
allocator_resident:9125888
allocator_muzzy:0
total_system_memory:66516324352
total_system_memory_human:61.95G
used_memory_lua:31744
used_memory_vm_eval:31744
used_memory_lua_human:31.00K
used_memory_scripts_eval:0
number_of_cached_scripts:0
number_of_functions:0
number_of_libraries:0
used_memory_vm_functions:33792
used_memory_vm_total:65536
used_memory_vm_total_human:64.00K
used_memory_functions:192
used_memory_scripts:192
used_memory_scripts_human:192B
maxmemory:0
maxmemory_human:0B
maxmemory_policy:noeviction
allocator_frag_ratio:1.39
allocator_frag_bytes:266352
allocator_rss_ratio:6.07
allocator_rss_bytes:7622656
rss_overhead_ratio:1.09
rss_overhead_bytes:835584
mem_fragmentation_ratio:15.13
mem_fragmentation_bytes:9303008
mem_not_counted_for_evict:0
mem_replication_backlog:0
mem_total_replication_buffers:0
mem_replica_full_sync_buffer:0
mem_clients_slaves:0
mem_clients_normal:0
mem_cluster_links:0
mem_aof_buffer:0
mem_allocator:jemalloc-5.3.0
mem_overhead_db_hashtable_rehashing:0
active_defrag_running:99
lazyfree_pending_objects:0
lazyfreed_objects:0

# Persistence
loading:0
async_loading:0
current_cow_peak:0
current_cow_size:0
current_cow_size_age:0
current_fork_perc:0.00
current_save_keys_processed:0
current_save_keys_total:0
rdb_changes_since_last_save:10
rdb_bgsave_in_progress:0
rdb_last_save_time:1756723699
rdb_last_bgsave_status:ok
rdb_last_bgsave_time_sec:-1
rdb_current_bgsave_time_sec:-1
rdb_saves:0
rdb_last_cow_size:0
rdb_last_load_keys_expired:0
rdb_last_load_keys_loaded:0
aof_enabled:0
aof_rewrite_in_progress:0
aof_rewrite_scheduled:0
aof_last_rewrite_time_sec:-1
aof_current_rewrite_time_sec:-1
aof_last_bgrewrite_status:ok
aof_rewrites:0
aof_rewrites_consecutive_failures:0
aof_last_write_status:ok
aof_last_cow_size:0
module_fork_in_progress:0
module_fork_last_cow_size:0

# Threads
io_thread_0:clients=1,reads=26,writes=24

# Stats
total_connections_received:2
total_commands_processed:23
instantaneous_ops_per_sec:0
total_net_input_bytes:1316
total_net_output_bytes:604
total_net_repl_input_bytes:0
total_net_repl_output_bytes:0
instantaneous_input_kbps:0.00
instantaneous_output_kbps:0.00
instantaneous_input_repl_kbps:0.00
instantaneous_output_repl_kbps:0.00
rejected_connections:0
sync_full:0
sync_partial_ok:0
sync_partial_err:0
expired_subkeys:0
expired_keys:0
expired_stale_perc:0.00
expired_time_cap_reached_count:0
expire_cycle_cpu_milliseconds:0
evicted_keys:0
evicted_clients:0
evicted_scripts:0
total_eviction_exceeded_time:0
current_eviction_exceeded_time:0
keyspace_hits:4
keyspace_misses:0
pubsub_channels:0
pubsub_patterns:0
pubsubshard_channels:0
latest_fork_usec:0
total_forks:0
migrate_cached_sockets:0
slave_expires_tracked_keys:0
active_defrag_hits:436
active_defrag_misses:0
active_defrag_key_hits:25
active_defrag_key_misses:0
total_active_defrag_time:49
current_active_defrag_time:1
tracking_total_keys:0
tracking_total_items:0
tracking_total_prefixes:0
unexpected_error_replies:0
total_error_replies:9
dump_payload_sanitizations:0
total_reads_processed:26
total_writes_processed:24
io_threaded_reads_processed:0
io_threaded_writes_processed:0
io_threaded_total_prefetch_batches:0
io_threaded_total_prefetch_entries:0
client_query_buffer_limit_disconnections:0
client_output_buffer_limit_disconnections:0
reply_buffer_shrinks:0
reply_buffer_expands:0
eventloop_cycles:12232
eventloop_duration_sum:112473
eventloop_duration_cmd_sum:45
instantaneous_eventloop_cycles_per_sec:6778
instantaneous_eventloop_duration_usec:0
acl_access_denied_auth:0
acl_access_denied_cmd:0
acl_access_denied_key:0
acl_access_denied_channel:0

# Replication
role:master
connected_slaves:0
master_failover_state:no-failover
master_replid:e72062577bca2b4b09e685e3f00dfa475f404c39
master_replid2:0000000000000000000000000000000000000000
master_repl_offset:0
second_repl_offset:-1
repl_backlog_active:0
repl_backlog_size:1048576
repl_backlog_first_byte_offset:0
repl_backlog_histlen:0

# CPU
used_cpu_sys:0.006058
used_cpu_user:0.112088
used_cpu_sys_children:0.000000
used_cpu_user_children:0.000000
used_cpu_sys_main_thread:0.006045
used_cpu_user_main_thread:0.111845

# Modules
module:name=vectorset,ver=1,api=1,filters=0,usedby=[],using=[],options=[handle-io-errors|handle-repl-async-load]

# Commandstats
cmdstat_ping:calls=1,usec=1,usec_per_call=1.00,rejected_calls=0,failed_calls=0
cmdstat_xreadgroup:calls=2,usec=14,usec_per_call=7.00,rejected_calls=0,failed_calls=0
cmdstat_del:calls=3,usec=2,usec_per_call=0.67,rejected_calls=0,failed_calls=0
cmdstat_xgroup|create:calls=3,usec=5,usec_per_call=1.67,rejected_calls=0,failed_calls=0
cmdstat_xdelex:calls=10,usec=6,usec_per_call=0.60,rejected_calls=1,failed_calls=8
cmdstat_select:calls=1,usec=1,usec_per_call=1.00,rejected_calls=0,failed_calls=0
cmdstat_xadd:calls=3,usec=16,usec_per_call=5.33,rejected_calls=0,failed_calls=0

# Errorstats
errorstat_ERR:count=9

# Latencystats
latency_percentiles_usec_ping:p50=1.003,p99=1.003,p99.9=1.003
latency_percentiles_usec_xreadgroup:p50=4.015,p99=10.047,p99.9=10.047
latency_percentiles_usec_del:p50=1.003,p99=1.003,p99.9=1.003
latency_percentiles_usec_xgroup|create:p50=1.003,p99=3.007,p99.9=3.007
latency_percentiles_usec_xdelex:p50=0.001,p99=2.007,p99.9=2.007
latency_percentiles_usec_select:p50=1.003,p99=1.003,p99.9=1.003
latency_percentiles_usec_xadd:p50=4.015,p99=11.007,p99.9=11.007

# Cluster
cluster_enabled:0

# Keyspace
db9:keys=2,expires=0,avg_ttl=0,subexpiry=0

# Keysizes

------ CLIENT LIST OUTPUT ------
id=5 addr=127.0.0.1:35317 laddr=127.0.0.1:21123 fd=12 name= age=0 idle=0 flags=N db=9 sub=0 psub=0 ssub=0 multi=-1 watch=0 qbuf=76 qbuf-free=20398 argv-mem=30 multi-mem=0 rbs=16384 rbp=16384 obl=4 oll=0 omem=0 tot-mem=37846 events=r cmd=xdelex user=default redir=-1 resp=2 lib-name= lib-ver= io-thread=0 tot-net-in=1309 tot-net-out=597 tot-cmds=22

------ CURRENT CLIENT INFO ------
id=5 addr=127.0.0.1:35317 laddr=127.0.0.1:21123 fd=12 name= age=0 idle=0 flags=N db=9 sub=0 psub=0 ssub=0 multi=-1 watch=0 qbuf=76 qbuf-free=20398 argv-mem=30 multi-mem=0 rbs=16384 rbp=16384 obl=4 oll=0 omem=0 tot-mem=37846 events=r cmd=xdelex user=default redir=-1 resp=2 lib-name= lib-ver= io-thread=0 tot-net-in=1309 tot-net-out=597 tot-cmds=22
argc: '7'
argv[0]: '"XDELEX"'
argv[1]: '"mystream"'
argv[2]: '"DELREF"'
argv[3]: '"IDS"'
argv[4]: '"2"'
argv[5]: '"1-0"'
argv[6]: '"2-0"'
117196:M 01 Sep 2025 18:48:19.911 # key 'mystream' found in DB containing the following object:
117196:M 01 Sep 2025 18:48:19.911 # Object type: 6
117196:M 01 Sep 2025 18:48:19.911 # Object encoding: 10
117196:M 01 Sep 2025 18:48:19.911 # Object refcount: 1

------ EXECUTING CLIENT INFO ------
id=5 addr=127.0.0.1:35317 laddr=127.0.0.1:21123 fd=12 name= age=0 idle=0 flags=N db=9 sub=0 psub=0 ssub=0 multi=-1 watch=0 qbuf=76 qbuf-free=20398 argv-mem=30 multi-mem=0 rbs=16384 rbp=16384 obl=4 oll=0 omem=0 tot-mem=37846 events=r cmd=xdelex user=default redir=-1 resp=2 lib-name= lib-ver= io-thread=0 tot-net-in=1309 tot-net-out=597 tot-cmds=22
argc: '7'
argv[0]: '"XDELEX"'
argv[1]: '"mystream"'
argv[2]: '"DELREF"'
argv[3]: '"IDS"'
argv[4]: '"2"'
argv[5]: '"1-0"'
argv[6]: '"2-0"'
117196:M 01 Sep 2025 18:48:19.911 # key 'mystream' found in DB containing the following object:
117196:M 01 Sep 2025 18:48:19.911 # Object type: 6
117196:M 01 Sep 2025 18:48:19.911 # Object encoding: 10
117196:M 01 Sep 2025 18:48:19.911 # Object refcount: 1

------ MODULES INFO OUTPUT ------

------ CONFIG DEBUG OUTPUT ------
activedefrag yes
slave-read-only yes
lazyfree-lazy-expire no
lazyfree-lazy-user-flush no
io-threads 1
sanitize-dump-payload no
repl-diskless-sync yes
proto-max-bulk-len 512mb
client-query-buffer-limit 1gb
lazyfree-lazy-server-del no
lazyfree-lazy-user-del no
replica-read-only yes
list-compress-depth 0
repl-diskless-load disabled
lazyfree-lazy-eviction no

------ FAST MEMORY TEST ------
117196:M 01 Sep 2025 18:48:19.911 # Bio worker thread #0 terminated
117196:M 01 Sep 2025 18:48:19.911 # Bio worker thread #1 terminated
117196:M 01 Sep 2025 18:48:19.911 # Bio worker thread #2 terminated
*** Preparing to test memory region 558e45eff000 (2322432 bytes)
*** Preparing to test memory region 558e51fed000 (135168 bytes)
*** Preparing to test memory region 7f059c000000 (135168 bytes)
*** Preparing to test memory region 7f05a2dfc000 (8388608 bytes)
*** Preparing to test memory region 7f05a35fd000 (8388608 bytes)
*** Preparing to test memory region 7f05a3dfe000 (8388608 bytes)
*** Preparing to test memory region 7f05a45ff000 (8388608 bytes)
*** Preparing to test memory region 7f05a4e00000 (8388608 bytes)
*** Preparing to test memory region 7f05a5600000 (10485760 bytes)
*** Preparing to test memory region 7f05a6600000 (8388608 bytes)
*** Preparing to test memory region 7f05a7005000 (53248 bytes)
*** Preparing to test memory region 7f05a747a000 (16384 bytes)
*** Preparing to test memory region 7f05a754d000 (28672 bytes)
*** Preparing to test memory region 7f05a7682000 (8192 bytes)
.O.O.O.O.O.O.O.O.O.O.O.O.O.O
Fast memory test PASSED, however your memory can still be broken. Please run a memory test for several hours if possible.

=== REDIS BUG REPORT END. Make sure to include from START to END. ===

       Please report the crash by opening an issue on github:

           http://github.com/redis/redis/issues

  If a Redis module was involved, please open in the module's repo instead.

  Suspect RAM error? Use redis-server --test-memory to verify it.

  Some other issues could be detected by redis-server --check-system
```